+++ pynmrstar/saveframe.py
@@ -2,6 +2,7 @@ import json
 import warnings
 from csv import reader as csv_reader, writer as csv_writer
 from io import StringIO
+from pathlib import Path
 from typing import TextIO, BinaryIO, Union, List, Optional, Any, Dict, Iterable, Tuple
 
 from pynmrstar import definitions, entry as entry_mod, loop as loop_mod, parser as parser_mod, utils
@@ -346,7 +347,7 @@ class Saveframe(object):
 
     @classmethod
     def from_file(cls,
-                  the_file: Union[str, TextIO, BinaryIO],
+                  the_file: Union[str, Path, TextIO, BinaryIO],
                   csv: bool = False,
                   convert_data_types: bool = False,
                   raise_parse_warnings: bool = False,
@@ -354,7 +355,8 @@ class Saveframe(object):
         """Create a saveframe by loading in a file. Specify csv=True is
         the file is a CSV file. If the_file starts with http://,
         https://, or ftp:// then we will use those protocols to attempt
-        to open the file.
+        to open the file. the_file can be a string path, pathlib.Path object,
+        or an open file handle.
 
         Setting convert_data_types to True will automatically convert
         the data loaded from the file into the corresponding python type as
@@ -998,7 +1000,7 @@ class Saveframe(object):
         return errors
 
     def write_to_file(self,
-                      file_name: str,
+                      file_name: Union[str, Path],
                       format_: str = "nmrstar",
                       show_comments: bool = True,
                       skip_empty_loops: bool = False,